+++ experiments/crbe_fixed_time_experiments.py
@@ -1,7 +1,7 @@
 # crbe_fixed_time_experiments.py
 import matplotlib.pyplot as plt
 import numpy as np
-import crbe
+from src import crbe # Changed import
 import meshio
 from tqdm import tqdm
 import time
