+++ experiments/optimal_hyperparams_search.py
@@ -2,8 +2,8 @@ import optuna
 import numpy as np
 import torch
 import time
-import pinn
-import crbe
+from src import pinn # Changed import
+from src import crbe # Changed import
 import meshio
 import os
 import time
